Add build results summary and raw output as exports
We export the build results, both the raw JSON response from AWS as well
as a nice human-friendly text string. As these are exports, they can be
used by later stages in the Bitrise build process, eg feeding into a
Slack step to return more information about how the build went.

Note that currently this does not include a break down of each
individual job inside the build run. This could be a useful extension
ontop of this work in future.